# openapi.json
{
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "Notification": {
        "properties": {
          "body": {
            "title": "Body",
            "type": "string"
          },
          "channel": {
            "title": "Channel",
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "orderId": {
            "title": "Orderid",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "userId": {
            "title": "Userid",
            "type": "string"
          }
        },
        "required": [
          "id",
          "userId",
          "orderId",
          "kind",
          "title",
          "body",
          "channel",
          "status",
          "createdAt"
        ],
        "title": "Notification",
        "type": "object"
      },
      "NotificationRequest": {
        "properties": {
          "deliveryWindow": {
            "anyOf": [
              {
                "enum": [
                  "morning",
                  "afternoon",
                  "evening"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": "morning",
            "title": "Deliverywindow"
          },
          "giftMessage": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Giftmessage"
          },
          "kind": {
            "const": "order_confirmed",
            "title": "Kind",
            "type": "string"
          },
          "orderId": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Orderid",
            "type": "string"
          },
          "totalCents": {
            "minimum": 0.0,
            "title": "Totalcents",
            "type": "integer"
          },
          "userId": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Userid",
            "type": "string"
          }
        },
        "required": [
          "userId",
          "orderId",
          "kind",
          "totalCents"
        ],
        "title": "NotificationRequest",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "wmd notification-service",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/v1/notifications": {
      "get": {
        "operationId": "list_for_user_v1_notifications_get",
        "parameters": [
          {
            "in": "query",
            "name": "userId",
            "required": true,
            "schema": {
              "maxLength": 64,
              "minLength": 1,
              "title": "Userid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Notification"
                  },
                  "title": "Response List For User V1 Notifications Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List For User"
      },
      "post": {
        "description": "Deliver once per order and kind: a retry returns the first notification (200).",
        "operationId": "send_v1_notifications_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NotificationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Notification"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Send"
      }
    }
  }
}